Made favourite files (seperated from user)
