+++ backend/index.js
@@ -5,6 +5,7 @@ import { router as searchRouter } from "./routers/searchRouter.js";
 import { router as movieRouter } from "./routers/movieRouter.js";
 import { router as userRouter } from "./routers/userRouter.js";
 import { router as reviewRouter } from "./routers/reviewRouter.js";
+import { router as favouriteRouter } from "./routers/favouriteRouter.js"
 
 dotenv.config();
 
@@ -18,6 +19,7 @@ app.use("/search", searchRouter);
 app.use("/movie", movieRouter);
 app.use("/user", userRouter);
 app.use("/review", reviewRouter);
+app.use("/favourite", favouriteRouter)
 
 app.use((err, req, res, next) => {
 	const statusCode = err.statusCode || 500;
